mnist/requirements.txt
numpy==1.19.5
PyYAML==6.0
torch==1.8.2
torchsummary==1.5.1
torchvision==0.9.2
tqdm==4.64.1
